Improve error messages and responses

+++ src/providers/providers.controller.ts
@@ -7,10 +7,14 @@ import {
     Body,
     Param,
     HttpStatus,
+    NotFoundException,
 } from '@nestjs/common';
 
 import { ProvidersService } from './providers.service';
-import { ProvidersDTO } from './dto/provider.dto';
+import { Provider } from './dto/provider.interface';
+import { CreateProviderDto } from './dto/create-provider.dto';
+import { validateOrReject } from 'class-validator';
+import { ProvidersEntity } from './provider.entity';
 
 @Controller('providers')
 export class ProvidersController {
@@ -27,40 +31,72 @@ export class ProvidersController {
     }
 
     @Post()
-    async createProviders(@Body() data: ProvidersDTO) {
-        const provider = await this.providersService.create(data);
-        return {
-            statusCode: HttpStatus.OK,
-            message: 'Provider created successfully',
-            provider
-        };
+    async createProvider(@Body() data: CreateProviderDto) {
+        try {
+            validateOrReject(data)
+            const provider = await this.providersService.create(data);
+            return {
+                statusCode: HttpStatus.OK,
+                message: 'Provider created successfully',
+                provider
+            };
+        } catch (errors) {
+            return {
+                statusCode: HttpStatus.BAD_REQUEST,
+                message: 'Caught promise rejection (validation failed).',
+                errors: errors
+            };
+        }
     }
 
     @Get(':id')
     async readProvider(@Param('id') id: number) {
-        const data = await this.providersService.showOne(id);
+        const provider: ProvidersEntity = await this.providersService.showOne(id);
+        if (provider === undefined) {
+            throw new NotFoundException('Invalid provider id');
+        }
         return {
             statusCode: HttpStatus.OK,
             message: 'Provider fetched successfully',
-            data,
+            provider,
         };
     }
 
     @Patch(':id')
-    async uppdateProvider(@Param('id') id: number, @Body() data: Partial<ProvidersDTO>) {
-        await this.providersService.update(id, data);
-        return {
-            statusCode: HttpStatus.OK,
-            message: 'Provider updated successfully',
-        };
+    async uppdateProvider(@Param('id') id: number, @Body() data: Partial<CreateProviderDto>) {
+        const provider: ProvidersEntity = await this.providersService.showOne(id);
+        if (provider === undefined) {
+            throw new NotFoundException('Invalid provider id');
+        }
+
+        try {
+            validateOrReject(data);
+            await this.providersService.update(id, data);
+            return {
+                statusCode: HttpStatus.OK,
+                message: 'Provider updated successfully',
+            };
+        } catch (errors) {
+            return {
+                statusCode: HttpStatus.BAD_REQUEST,
+                message: 'Caught promise rejection (validation failed).',
+                errors: errors
+            };
+        }
     }
 
     @Delete(':id')
     async deleteProvider(@Param('id') id: number) {
+        const provider: ProvidersEntity = await this.providersService.showOne(id);
+        if (provider === undefined) {
+            throw new NotFoundException('Invalid provider id');
+        }
+
         await this.providersService.remove(id);
         return {
             statusCode: HttpStatus.OK,
             message: 'Provider deleted successfully',
         };
+
     }
 }
